requirements.txt
appdirs==1.4.4
arrow==1.0.3
beautifulsoup4==4.9.3
boto3==1.18.58
botocore==1.21.58
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
filelock==3.0.12
googlesearch-python==1.1.0
idna==2.10
importlib-metadata==3.7.0
jmespath==0.10.0
lxml==4.6.3
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.5.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
urllib3==1.25.9
virtualenv==20.4.0
zipp==3.4.0
